ui-service/requirements.txt
requests
streamlit
pandas
python-dotenv==1.0.0
streamlit-chat==0.1.1
lxml==4.9.3